\name{init}
\alias{init}
\title{Initialize Java Virtual Machine}
\usage{
  init()
}
\description{
  Function \code{init} start the Java virtual machine, and
  add necessary class paths to the Java virtual machine.
  Note that if you load this package, the initialization is
  taken care of automatically and there is no necessary to
  call this function manually. Function \code{init} is for
  people who want to use this package without loading it.
}

